# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для проекта DoorDashers.",
        "title": "AdminAdmin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5458",
    "basePath": "/api",
    "paths": {
...
        "/cart": {
            "get": {
                "description": "Возвращает список товаров в корзине текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Получить корзину",
                "responses": {
                    "200": {
                        "description": "Успешное получение корзины",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CartItem"
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации (проблемы с куки или JWT)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/update/{productID}": {
            "post": {
                "description": "Обновляет количество товара в корзине для текущего пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Обновление количества продуктов в корзине",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID продукта",
                        "name": "productID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры для изменения количества товара",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CartUpdateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление количества товара в корзине"
                    },
                    "400": {
                        "description": "Некорректный формат данных",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неавторизован (проблемы с куки или JWT)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера при обновлении корзины",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/restaurants/list": {
            "get": {
                "description": "Получение списка ресторанов с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Список ресторанов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Количество элементов",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение списка ресторанов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Restaurant"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/restaurants/{id}": {
            "get": {
                "description": "Получение списка продуктов ресторана с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Получить продукты ресторана",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID ресторана",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество элементов (по умолчанию 100)",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение продуктов ресторана",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID ресторана",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.CartUpdateBody": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
...
        "models.CartItem": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "models.Restaurant": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}